Cast to long long before multiplication to avoid overflow
On 32-bit systems, time_t is 31 bits and overflows after ~24 days. This should
fix #38.

+++ licq/src/mainloop.cpp
@@ -98,7 +98,7 @@ long long MainLoop::Private::getMonotonicClock()
   // Get monotonic time and convert to milliseconds
   struct timespec ts;
   clock_gettime(CLOCK_MONOTONIC, &ts);
-  return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+  return static_cast<long long>(ts.tv_sec) * 1000 + ts.tv_nsec / 1000000;
 }
 
 MainLoop::MainLoop()
